+++ src/components/BreakpointDistance.tsx
@@ -1,5 +1,7 @@
 import React from 'react'
 import Separator from './Separator'
+import type { Language } from '../types'
+import { t } from '../utils/translations'
 
 interface BreakpointDistanceProps {
   distance: number
@@ -7,6 +9,7 @@ interface BreakpointDistanceProps {
   breakpointKey: string
   separatorColor: string
   textClass: string
+  language?: Language
 }
 
 const BreakpointDistance: React.FC<BreakpointDistanceProps> = ({
@@ -15,11 +18,12 @@ const BreakpointDistance: React.FC<BreakpointDistanceProps> = ({
   breakpointKey,
   separatorColor,
   textClass,
+  language = 'en',
 }: BreakpointDistanceProps) => (
   <>
     <Separator color={separatorColor} />
     <span className={textClass}>
-      {`${prefix}${distance}px to ${breakpointKey}`}
+      {`${prefix}${distance}px ${t('breakpointDistance', language)} ${breakpointKey}`}
     </span>
   </>
 )