+++ src/index.ts
@@ -1,1 +1,9 @@
+// Components
 export { default as ScreenSizeOverlay } from './components/ScreenSizeOverlay'
+
+// Themes
+export * from './themes'
+
+// Hooks
+export { useWindowSize } from './hooks/useWindowSize'
+export { useBreakpointInfo } from './hooks/useBreakpointInfo'